+++ R/sample.R
@@ -61,7 +61,9 @@ st_sample = function(x, size, ..., type = "random") {
 	}
 }
 
-st_poly_sample = function(x, size, ..., type = "random", offset = st_sample(st_as_sfc(st_bbox(x)), 1)[[1]]) {
+st_poly_sample = function(x, size, ..., type = "random", 
+		offset = st_sample(st_as_sfc(st_bbox(x)), 1)[[1]]) {
+
 	a0 = st_area(st_make_grid(x, n = c(1,1)))
 	a1 = sum(st_area(x))
 	# st_polygon(list(rbind(c(-180,-90),c(180,-90),c(180,90),c(-180,90),c(-180,-90))))
@@ -77,8 +79,9 @@ st_poly_sample = function(x, size, ..., type = "random", offset = st_sample(st_a
 		dx = sqrt(a0 / size / (sqrt(3)/2))
 		hex_grid(x, pt = offset, dx = dx, points = TRUE, clip = FALSE)
 	} else if (type == "regular") {
-		dx = sqrt(a0 / size)
-		offset = c((offset[1] - bb["xmin"]) %% dx, (offset[2] - bb["ymin"]) %% dx) + bb[c("xmin", "ymin")]
+		dx = as.numeric(sqrt(a0 / size))
+		offset = c((offset[1] - bb["xmin"]) %% dx, 
+			(offset[2] - bb["ymin"]) %% dx) + bb[c("xmin", "ymin")]
 		n = c(round((bb["xmax"] - offset[1])/dx), round((bb["ymax"] - offset[2])/dx))
 		st_make_grid(x, cellsize = c(dx, dx), offset = offset, n = n, what = "corners")
 	} else if (type == "random") {
@@ -131,8 +134,8 @@ st_ll_sample = function (x, size, ..., type = "random", offset = runif(1)) {
 ## - contains pt
 ## - has x spacing dx: the shortest distance between x coordinates with identical y coordinate
 ## - selects geometries intersecting with obj
-hex_grid = function(obj, pt = bb[c("xmin", "ymin")], dx = diff(st_bbox(obj)[c("xmin", "xmax")])/10.1, 
-	points = TRUE, clip = NA) {
+hex_grid = function(obj, pt = bb[c("xmin", "ymin")], 
+		dx = diff(st_bbox(obj)[c("xmin", "xmax")])/10.1, points = TRUE, clip = NA) {
 
 	bb = st_bbox(obj)
 	dy = sqrt(3) * dx / 2
@@ -140,7 +143,7 @@ hex_grid = function(obj, pt = bb[c("xmin", "ymin")], dx = diff(st_bbox(obj)[c("x
 	ylim = bb[c("ymin", "ymax")]
 	offset = c(x = (pt[1] - xlim[1]) %% dx, y = (pt[2] - ylim[1]) %% (2 * dy))
 	x = seq(xlim[1] - dx, xlim[2] + dx, dx) + offset[1]
-	y = seq(ylim[1] - 2 * dy, ylim[2] + dy, dy) + offset[2]
+	y = seq(ylim[1] - 2 * dy, ylim[2] + 2 * dy, dy) + offset[2]
 
 	y <- rep(y, each = length(x))
 	x <- rep(c(x, x + dx / 2), length.out = length(y))
@@ -152,16 +155,16 @@ hex_grid = function(obj, pt = bb[c("xmin", "ymin")], dx = diff(st_bbox(obj)[c("x
 		dx2 = dx / 2
 		x.offset = c(-dx / 2, 0, dx / 2, dx / 2, 0, -dx / 2, -dx / 2)
 		y.offset = c(dy / 2, dy, dy / 2, -dy / 2, -dy, -dy / 2, dy / 2)
-		xy = cbind(x, y)[x >= xlim[1] - dx2 & x <= xlim[2] + dx2 & y >= ylim[1] - dx2 & y <= ylim[2] + dx2, ]
+		xy = cbind(x, y)[x >= xlim[1] - dx2 & x <= xlim[2] + dx2 & y >= ylim[1] - dy & y <= ylim[2] + dy, ]
 		mk_pol = function(pt) { st_polygon(list(cbind(pt[1] + x.offset, pt[2] + y.offset))) }
 		st_sfc(lapply(seq_len(nrow(xy)), function(i) mk_pol(xy[i,])), crs = st_crs(bb))
 	}
-	i = if (clip) {
+	sel = if (isTRUE(clip)) {
 		if (points)
 			lengths(st_intersects(ret, obj)) > 0
 		else
 			lengths(st_relate(ret, obj, "2********")) > 0
 	} else
 		TRUE
-	ret[i]
+	ret[sel]
 }